refactor(Page): move styles
